+++ SwipeView/SwipeView.h
@@ -1,7 +1,7 @@
 //
 //  SwipeView.h
 //
-//  Version 1.1.5
+//  Version 1.1.6
 //
 //  Created by Nick Lockwood on 03/09/2010.
 //  Copyright 2010 Charcoal Design